+++ neo4j/_sync/driver.py
@@ -94,10 +94,17 @@ class GraphDatabase:
 
         if driver_type == DRIVER_BOLT:
             if parse_routing_context(parsed.query):
-                raise ValueError(
-                    'Routing parameters are not supported with scheme "bolt". '
-                    'Given URI "{}".'.format(uri)
+                deprecation_warn(
+                    "Creating a direct driver (`bolt://` scheme) with routing "
+                    "context (URI parameters) is deprecated. They will be "
+                    "ignored. This will raise an error in a future release. "
+                    'Given URI "{}"'.format(uri)
                 )
+                # TODO: 6.0 - raise instead of warning
+                # raise ValueError(
+                #     'Routing parameters are not supported with scheme '
+                #     '"bolt". Given URI "{}".'.format(uri)
+                # )
             return cls.bolt_driver(parsed.netloc, auth=auth, **config)
         elif driver_type == DRIVER_NEO4j:
             routing_context = parse_routing_context(parsed.query)